requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.32.1
langgraph==0.2.74
langchain-core==0.3.25
langchain-openai==0.2.10
pydantic==2.10.3
python-multipart==0.0.12
jinja2==3.1.4
aiofiles==24.1.0
pandas==2.2.3
python-dotenv==1.0.1
